Bugfix - results not being stored in the database table.
The previous fix did not work - so I've added actual tests, where the task is reloaded from the database and we can check that the results are actually present in the database record.

Also I obviously did not understand how `composed_of` works, so I've removed it.  This means that results is an OpenStruct in your handlers, but it's a Hash when you access it from the task itself.