Is this task a zombie?
If this is a background task and it has not been updated for more than 3 times the delay, it's possible that the background task-runner job has died.  So, the task records a "becomes_zombie_time" value in the database, which can then be used to identify zombie tasks and restart them.