Sleep times - immediate vs later
When a task is started in the background, the very first "wake time"
will be immediately; subsequent sleeps and wakes will use the setting
defined by `delay`.

This means that you can set a large delay but still have the very first
set of handlers run not long after you `start`ed the task.

Previously if you started a background task and you wanted it to respond
quickly you would need to reduce the `delay` to a low value resulting in
lots of spinning if actually subsequent handlers could wait for a long
time